Decimalize the repeatable IP multiplier upgrade
This is useful if mods want to uncap the repeatable IP multiplier upgrade.

If the endgame is extended to F9e15 antimatter, nine times out of ten, the repeatable IP multiplier upgrade will be able to be purchased more than e308 times, even if its cost scales faster with more purchases.