Fix GameEnd state determination
This brings the calculation of GameEnd.endState back in line with the vanilla game, bringing back tab zalgoification in the process.

The endState calculation was previously changed to check if the layer count of the total antimatter was at least 2, which meant endState would jump from 0 to at least 1 as soon as total Antimatter hit e9e15, bypassing not only tab zalgoification but also (depending on Antimatter count) other phases of the end sequence.