Fix #23: Tesseract effect was amplified over time
Due to the way the Infinity Dimensions tab initialized its variables, staying on the tab and maxing out the number of purchases for an Infinity Dimension caused the base purchase cap to increase by the bonus from Tesseracts.

The net cause was a reactivity error that was not directly caught in my testing setup.

I've fixed this by initializing the relevant variable in a way that doesn't link or induce reactivity.